refactor: move RPC help definitions out of core_io module into rpc/
The GetJsonHelp()/GetRpcResult() definitions lived in the {evo,llmq,
governance}/core_write.cpp files, which form the core_io module and are built
into libbitcoin_common for the sake of their ToJson() halves (dash-tx prints
special-tx payloads via TxToUniv()). The help halves forced rpc/util.h into
core_io's dependency closure, creating four circular dependencies through
core_io -> rpc/util -> node/transaction, and shipped the help tables in
libbitcoin_common although every consumer (rpc/blockchain, rpc/coinjoin,
rpc/evo, rpc/governance, rpc/masternode, rpc/quorums, rpc/rawtransaction) is
in libbitcoin_node.

Move RPCRESULT_MAP, GetRpcResult() and all GetJsonHelp() definitions verbatim
into a new rpc/json_help.{h,cpp} in libbitcoin_node, and move the
GetRpcResult() declaration from core_io.h to the new header. The core_write
files keep only their ToJson() definitions. The help is RPC documentation
shared by several command files, so it gets its own translation unit rather
than being spliced into the generic machinery of rpc/util.cpp, which is part
of libbitcoin_common and would keep the tables in the common library.

The four coinjoin suppressions through core_io -> rpc/util are gone; the
longer pre-existing coinjoin/client -> coinjoin/util -> wallet/wallet cycle
they had been shadowing is visible to the linter again and returns to the
suppression list.
