+++ src/core_io.h
@@ -18,7 +18,6 @@ class CTxUndo;
 class uint256;
 struct CMutableTransaction;
 struct CSpentIndexTxInfo;
-struct RPCResult;
 
 class UniValue;
 
@@ -56,8 +55,5 @@ std::string EncodeHexTx(const CTransaction& tx);
 std::string SighashToStr(unsigned char sighash_type);
 void ScriptToUniv(const CScript& script, UniValue& out, bool include_hex = true, bool include_address = false);
 void TxToUniv(const CTransaction& tx, const uint256& block_hash, UniValue& entry, bool include_hex = true, int serialize_flags = 0, const CTxUndo* txundo = nullptr, TxVerbosity verbosity = TxVerbosity::SHOW_DETAILS, const CSpentIndexTxInfo* ptxSpentInfo = nullptr);
-
-// evo/core_write.cpp
-RPCResult GetRpcResult(const std::string& key, bool optional = false, const std::string& override_name = "");
 
 #endif // BITCOIN_CORE_IO_H
