+++ src/governance/core_write.cpp
@@ -2,28 +2,12 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <core_io.h>
 #include <governance/common.h>
 #include <governance/governance.h>
 #include <governance/object.h>
-#include <rpc/util.h>
-#include <tinyformat.h>
 #include <util/check.h>
 
 #include <univalue.h>
-
-RPCResult CGovernanceManager::GetJsonHelp(const std::string& key, bool optional)
-{
-    return {RPCResult::Type::OBJ, key, optional, key.empty() ? "" : "Count of governance objects and votes",
-    {
-        {RPCResult::Type::NUM, "objects_total", "Total number of all governance objects"},
-        {RPCResult::Type::NUM, "proposals", "Number of governance proposals"},
-        {RPCResult::Type::NUM, "triggers", "Number of triggers"},
-        {RPCResult::Type::NUM, "other", "Total number of unknown governance objects"},
-        {RPCResult::Type::NUM, "erased", "Number of removed (expired) objects"},
-        {RPCResult::Type::NUM, "votes", "Total number of votes"},
-    }};
-}
 
 UniValue CGovernanceManager::ToJson() const
 {
@@ -57,46 +41,9 @@ UniValue CGovernanceManager::ToJson() const
     return jsonObj;
 }
 
-RPCResult CGovernanceObject::GetInnerJsonHelp(const std::string& key, bool optional)
-{
-    return Governance::Object::GetJsonHelp(key, optional);
-}
-
 UniValue CGovernanceObject::GetInnerJson() const
 {
     return m_obj.ToJson();
-}
-
-// CGovernanceObject::GetStateJson() defined in governance/object.cpp
-RPCResult CGovernanceObject::GetStateJsonHelp(const std::string& key, bool optional, const std::string& local_valid_key)
-{
-    return {RPCResult::Type::OBJ, key, optional, key.empty() ? "" : "Object state info",
-    {
-        {RPCResult::Type::STR_HEX, "DataHex", "Governance object (hex)"},
-        {RPCResult::Type::STR, "DataString", "Governance object (string)"},
-        {RPCResult::Type::STR_HEX, "Hash", "Hash of governance object"},
-        GetRpcResult("collateralHash", /*optional=*/false, /*override_name=*/"CollateralHash"),
-        {RPCResult::Type::NUM, "ObjectType", "Object types"},
-        {RPCResult::Type::NUM, "CreationTime", "Object creation timestamp"},
-        {RPCResult::Type::STR_HEX, "SigningMasternode", /*optional=*/true, "Signing masternode’s vin (for triggers only)"},
-        {RPCResult::Type::BOOL, local_valid_key, "Returns true if valid"},
-        {RPCResult::Type::STR, "IsValidReason", strprintf("%s error (human-readable string, empty if %s true)", local_valid_key, local_valid_key)},
-        {RPCResult::Type::BOOL, "fCachedValid", "Returns true if minimum support has been reached flagging object as valid"},
-        {RPCResult::Type::BOOL, "fCachedFunding", "Returns true if minimum support has been reached flagging object as fundable"},
-        {RPCResult::Type::BOOL, "fCachedDelete", "Returns true if minimum support has been reached flagging object as marked for deletion"},
-        {RPCResult::Type::BOOL, "fCachedEndorsed", "Returns true if minimum support has been reached flagging object as endorsed"},
-    }};
-}
-
-RPCResult CGovernanceObject::GetVotesJsonHelp(const std::string& key, bool optional)
-{
-    return {RPCResult::Type::OBJ, key, optional, key.empty() ? "" : "Object vote counts",
-    {
-        {RPCResult::Type::NUM, "AbsoluteYesCount", "Number of Yes votes minus number of No votes"},
-        {RPCResult::Type::NUM, "YesCount", "Number of Yes votes"},
-        {RPCResult::Type::NUM, "NoCount", "Number of No votes"},
-        {RPCResult::Type::NUM, "AbstainCount", "Number of Abstain votes"},
-    }};
 }
 
 UniValue CGovernanceObject::GetVotesJson(const CDeterministicMNList& tip_mn_list, vote_signal_enum_t signal) const
@@ -110,35 +57,6 @@ UniValue CGovernanceObject::GetVotesJson(const CDeterministicMNList& tip_mn_list
 }
 
 namespace Governance {
-RPCResult Object::GetJsonHelp(const std::string& key, bool optional)
-{
-    return {RPCResult::Type::OBJ, key, optional, key.empty() ? "" : "Object info",
-    {
-        {RPCResult::Type::STR_HEX, "objectHash", "Hash of proposal object"},
-        {RPCResult::Type::STR_HEX, "parentHash", "Hash of the parent object (root node has a hash of 0)"},
-        GetRpcResult("collateralHash"),
-        {RPCResult::Type::NUM, "createdAt", "Proposal creation timestamp"},
-        {RPCResult::Type::NUM, "revision", "Proposal revision number"},
-        // Everything but "hex" is the submitter's own JSON echoed back verbatim, so neither the
-        // set of keys nor their types is ours to promise: a proposal may carry extra fields and
-        // may encode the numeric ones as strings. Hence the type check is skipped here.
-        {RPCResult::Type::OBJ, "data", "", {
-            // Fields emitted through GetDataAsPlainString(), read by CProposalValidator
-            {RPCResult::Type::ANY, "end_epoch", /*optional=*/true, "Proposal end timestamp"},
-            {RPCResult::Type::ANY, "name", /*optional=*/true, "Proposal name"},
-            {RPCResult::Type::ANY, "payment_address", /*optional=*/true, "Proposal payment address"},
-            {RPCResult::Type::ANY, "payment_amount", /*optional=*/true, "Proposal payment amount"},
-            {RPCResult::Type::ANY, "start_epoch", /*optional=*/true, "Proposal start timestamp"},
-            {RPCResult::Type::ANY, "type", /*optional=*/true, "Object type"},
-            {RPCResult::Type::ANY, "url", /*optional=*/true, "Proposal URL"},
-            // Failure case for GetDataAsPlainString()
-            {RPCResult::Type::STR, "plain", /*optional=*/true, "Governance object data as string"},
-            // Always emitted by ToJson()
-            {RPCResult::Type::STR_HEX, "hex", "Governance object data as hex"},
-        }, /*skip_type_check=*/true},
-    }};
-}
-
 UniValue Object::ToJson() const
 {
     UniValue obj(UniValue::VOBJ);
