+++ src/rpc/masternode.cpp
@@ -10,6 +10,7 @@
 #include <governance/superblock.h>
 #include <masternode/payments.h>
 #include <rpc/evo_util.h>
+#include <rpc/json_help.h>
 
 #include <chainparams.h>
 #include <core_io.h>
