+++ src/rpc/quorums.cpp
@@ -22,6 +22,7 @@
 #include <llmq/snapshot.h>
 #include <llmq/utils.h>
 #include <rpc/evo_util.h>
+#include <rpc/json_help.h>
 #include <util/helpers.h>
 
 #include <chainparams.h>
