+++ test/util/data/non-backported.txt
@@ -54,6 +54,7 @@ src/rpc/coinjoin.cpp
 src/rpc/evo.cpp
 src/rpc/evo_util.*
 src/rpc/governance.cpp
+src/rpc/json_help.*
 src/rpc/masternode.cpp
 src/rpc/quorums.cpp
 src/saltedhasher.*
